+++ src/sparkline/index.js
@@ -1,0 +1,1 @@
+export { RowanSparkline } from "./sparkline.js";