+++ src/sparkline/sparkline.d.ts
@@ -1,0 +1,3 @@
+import "../elements.js";
+
+export { RowanSparkline } from "../../types/sparkline/sparkline.js";